Replace ints with size_t

+++ src/common.h
@@ -6,8 +6,8 @@ struct vector {
     int y;
 };
 
-static const int SCREEN_WIDTH = 640;
-static const int SCREEN_HEIGHT = 480;
-static const int GRID_WIDTH = SCREEN_WIDTH/2;
-static const int GRID_HEIGHT = SCREEN_HEIGHT;
-static const int GRID_SQUARE_LENGTH = 24;
+static const size_t SCREEN_WIDTH = 640;
+static const size_t SCREEN_HEIGHT = 480;
+static const size_t GRID_WIDTH = SCREEN_WIDTH/2;
+static const size_t GRID_HEIGHT = SCREEN_HEIGHT;
+static const size_t GRID_SQUARE_LENGTH = 24;
